+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ---
 
+## [**1.0.5**] - 2021-03-13
+
+### Added
+
+* Checks to the shell script (check if the files exist)
+* Colorized the output of the script.
+
 ## [**1.0.4**] - 2021-02-26
 
 ### Added